+++ src/components/ApplicationViewContainer/resources/financial-details/liabilities/credit/CreditLiabilityCard.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 import { Box } from '@mui/material';
 import { STRING_HELPERS } from '../../../../../../utils/string-helpers';
 import { AttributeComponent } from '../AttributeComponent';
