+++ src/components/PlaidLinkWidget/index.jsx
@@ -1,3 +1,2 @@
-
 import PlaidLinkContainer from './PlaidLinkContainer';
 export { PlaidLinkContainer };
